Update log keyword

+++ monitor-server/services/db/log_keyword.go
@@ -7,37 +7,41 @@ import (
 	"time"
 )
 
-func GetLogKeywordByServiceGroup(serviceGroupGuid string) (result []*models.LogKeywordServiceGroupObj,err error) {
-	serviceGroupObj,getErr := getSimpleServiceGroup(serviceGroupGuid)
+func GetLogKeywordByServiceGroup(serviceGroupGuid string) (result []*models.LogKeywordServiceGroupObj, err error) {
+	serviceGroupObj, getErr := getSimpleServiceGroup(serviceGroupGuid)
 	if getErr != nil {
-		return result,getErr
+		return result, getErr
 	}
 	result = []*models.LogKeywordServiceGroupObj{}
 	var logKeywordTable []*models.LogKeywordMonitorTable
 	err = x.SQL("select * from log_keyword_monitor where service_group=?", serviceGroupGuid).Find(&logKeywordTable)
 	if err != nil {
-		return result,fmt.Errorf("Query table fail,%s ", err.Error())
+		return result, fmt.Errorf("Query table fail,%s ", err.Error())
 	}
 	configList := []*models.LogKeywordMonitorObj{}
-	for _,v := range logKeywordTable {
-		configObj := models.LogKeywordMonitorObj{Guid: v.Guid,ServiceGroup: serviceGroupGuid,LogPath: v.LogPath,MonitorType: v.MonitorType}
+	for _, v := range logKeywordTable {
+		configObj := models.LogKeywordMonitorObj{Guid: v.Guid, ServiceGroup: serviceGroupGuid, LogPath: v.LogPath, MonitorType: v.MonitorType}
 		configObj.KeywordList = ListLogKeyword(v.Guid)
 		configObj.EndpointRel = ListLogKeywordEndpointRel(v.Guid)
 		configList = append(configList, &configObj)
 	}
-	result = append(result, &models.LogKeywordServiceGroupObj{ServiceGroupTable:serviceGroupObj,Config: configList})
+	result = append(result, &models.LogKeywordServiceGroupObj{ServiceGroupTable: serviceGroupObj, Config: configList})
 	return
 }
 
-func GetLogKeywordByEndpoint(endpointGuid string) (result []*models.LogKeywordServiceGroupObj,err error) {
+func GetLogKeywordByEndpoint(endpointGuid string, onlySource bool) (result []*models.LogKeywordServiceGroupObj, err error) {
 	result = []*models.LogKeywordServiceGroupObj{}
 	var logKeywordMonitorTable []*models.LogKeywordMonitorTable
-	err = x.SQL("select distinct t2.service_group from log_keyword_endpoint_rel t1 left join log_keyword_monitor t2 on t1.log_keyword_monitor=t2.guid where t1.source_endpoint=? or t1.target_endpoint=?", endpointGuid, endpointGuid).Find(&logKeywordMonitorTable)
-	if err != nil {
-		return result,fmt.Errorf("Query table fail,%s ", err.Error())
+	if onlySource {
+		err = x.SQL("select distinct t2.service_group from log_keyword_endpoint_rel t1 left join log_keyword_monitor t2 on t1.log_keyword_monitor=t2.guid where t1.source_endpoint=?", endpointGuid).Find(&logKeywordMonitorTable)
+	} else {
+		err = x.SQL("select distinct t2.service_group from log_keyword_endpoint_rel t1 left join log_keyword_monitor t2 on t1.log_keyword_monitor=t2.guid where t1.source_endpoint=? or t1.target_endpoint=?", endpointGuid, endpointGuid).Find(&logKeywordMonitorTable)
 	}
-	for _,v := range logKeywordMonitorTable {
-		tmpResult,tmpErr := GetLogKeywordByServiceGroup(v.ServiceGroup)
+	if err != nil {
+		return result, fmt.Errorf("Query table fail,%s ", err.Error())
+	}
+	for _, v := range logKeywordMonitorTable {
+		tmpResult, tmpErr := GetLogKeywordByServiceGroup(v.ServiceGroup)
 		if tmpErr != nil {
 			err = tmpErr
 			break
@@ -52,9 +56,9 @@ func CreateLogKeywordMonitor(param *models.LogKeywordMonitorCreateObj) (err erro
 	var actions []*Action
 	nowTime := time.Now().Format(models.DatetimeFormat)
 	logKeywordGuidList := guid.CreateGuidList(len(param.LogPath))
-	for i,path := range param.LogPath {
-		actions = append(actions, &Action{Sql: "insert into log_keyword_monitor(guid,service_group,log_path,monitor_type,update_time) value (?,?,?,?,?)",Param: []interface{}{logKeywordGuidList[i],param.ServiceGroup,path,param.MonitorType,nowTime}})
-		endpointRelActions,tmpErr := getLogKeywordEndpointRelCreateAction(param.EndpointRel, logKeywordGuidList[i])
+	for i, path := range param.LogPath {
+		actions = append(actions, &Action{Sql: "insert into log_keyword_monitor(guid,service_group,log_path,monitor_type,update_time) value (?,?,?,?,?)", Param: []interface{}{logKeywordGuidList[i], param.ServiceGroup, path, param.MonitorType, nowTime}})
+		endpointRelActions, tmpErr := getLogKeywordEndpointRelCreateAction(param.EndpointRel, logKeywordGuidList[i])
 		if tmpErr != nil {
 			err = tmpErr
 			break
@@ -67,14 +71,14 @@ func CreateLogKeywordMonitor(param *models.LogKeywordMonitorCreateObj) (err erro
 	return Transaction(actions)
 }
 
-func getLogKeywordEndpointRelCreateAction(param []*models.LogKeywordEndpointRelTable,logKeywordMonitor string) (actions []*Action,err error) {
+func getLogKeywordEndpointRelCreateAction(param []*models.LogKeywordEndpointRelTable, logKeywordMonitor string) (actions []*Action, err error) {
 	endpointRelGuidList := guid.CreateGuidList(len(param))
-	for ii,endpointRel := range param {
+	for ii, endpointRel := range param {
 		if endpointRel.SourceEndpoint == "" || endpointRel.TargetEndpoint == "" {
 			err = fmt.Errorf("endpointRel source and target can not empty ")
 			break
 		}
-		actions = append(actions, &Action{Sql: "insert into log_keyword_endpoint_rel(guid,log_keyword_monitor,source_endpoint,target_endpoint) value (?,?,?,?)",Param: []interface{}{endpointRelGuidList[ii],logKeywordMonitor,endpointRel.SourceEndpoint,endpointRel.TargetEndpoint}})
+		actions = append(actions, &Action{Sql: "insert into log_keyword_endpoint_rel(guid,log_keyword_monitor,source_endpoint,target_endpoint) value (?,?,?,?)", Param: []interface{}{endpointRelGuidList[ii], logKeywordMonitor, endpointRel.SourceEndpoint, endpointRel.TargetEndpoint}})
 	}
 	return
 }
@@ -82,9 +86,9 @@ func getLogKeywordEndpointRelCreateAction(param []*models.LogKeywordEndpointRelT
 func UpdateLogKeywordMonitor(param *models.LogKeywordMonitorObj) (err error) {
 	var actions []*Action
 	nowTime := time.Now().Format(models.DatetimeFormat)
-	actions = append(actions, &Action{Sql: "update log_keyword_monitor set log_path=?,monitor_type=?,update_time=? where guid=?",Param: []interface{}{param.LogPath,param.MonitorType,nowTime,param.Guid}})
-	actions = append(actions, &Action{Sql: "delete from log_keyword_endpoint_rel where log_keyword_monitor=?",Param: []interface{}{param.Guid}})
-	endpointRelActions,tmpErr := getLogKeywordEndpointRelCreateAction(param.EndpointRel, param.Guid)
+	actions = append(actions, &Action{Sql: "update log_keyword_monitor set log_path=?,monitor_type=?,update_time=? where guid=?", Param: []interface{}{param.LogPath, param.MonitorType, nowTime, param.Guid}})
+	actions = append(actions, &Action{Sql: "delete from log_keyword_endpoint_rel where log_keyword_monitor=?", Param: []interface{}{param.Guid}})
+	endpointRelActions, tmpErr := getLogKeywordEndpointRelCreateAction(param.EndpointRel, param.Guid)
 	if tmpErr != nil {
 		err = tmpErr
 		return
@@ -94,11 +98,25 @@ func UpdateLogKeywordMonitor(param *models.LogKeywordMonitorObj) (err error) {
 }
 
 func DeleteLogKeywordMonitor(logKeywordMonitorGuid string) (err error) {
+	var logKeywordMonitorTable []*models.LogKeywordMonitorTable
+	err = x.SQL("select * from log_keyword_monitor where guid=?", logKeywordMonitorGuid).Find(&logKeywordMonitorTable)
+	if len(logKeywordMonitorTable) == 0 {
+		return
+	}
+	var hostEndpointList []string
+	for _, v := range ListLogKeywordEndpointRel(logKeywordMonitorGuid) {
+		hostEndpointList = append(hostEndpointList, v.SourceEndpoint)
+	}
 	var actions []*Action
-	actions = append(actions, &Action{Sql: "delete from log_keyword_endpoint_rel where log_keyword_monitor=?",Param: []interface{}{logKeywordMonitorGuid}})
-	actions = append(actions, &Action{Sql: "delete from log_keyword_config where log_keyword_monitor=?",Param: []interface{}{logKeywordMonitorGuid}})
-	actions = append(actions, &Action{Sql: "delete from log_keyword_monitor where guid=?",Param: []interface{}{logKeywordMonitorGuid}})
-	return Transaction(actions)
+	actions = append(actions, &Action{Sql: "delete from log_keyword_endpoint_rel where log_keyword_monitor=?", Param: []interface{}{logKeywordMonitorGuid}})
+	actions = append(actions, &Action{Sql: "delete from log_keyword_config where log_keyword_monitor=?", Param: []interface{}{logKeywordMonitorGuid}})
+	actions = append(actions, &Action{Sql: "delete from log_keyword_monitor where guid=?", Param: []interface{}{logKeywordMonitorGuid}})
+	err = Transaction(actions)
+	if err != nil {
+		return err
+	}
+	err = SyncLogKeywordExporterConfig(hostEndpointList)
+	return
 }
 
 func ListLogKeyword(logKeywordMonitor string) (result []*models.LogKeywordConfigTable) {
@@ -108,17 +126,17 @@ func ListLogKeyword(logKeywordMonitor string) (result []*models.LogKeywordConfig
 }
 
 func CreateLogKeyword(param *models.LogKeywordConfigTable) (err error) {
-	_,err = x.Exec("insert into log_keyword_config(guid,log_keyword_monitor,keyword,regulative,notify_enable,priority,update_time) value (?,?,?,?,?,?,?)",guid.CreateGuid(),param.LogKeywordMonitor,param.Keyword,param.Regulative,param.NotifyEnable,param.Priority,time.Now().Format(models.DatetimeFormat))
+	_, err = x.Exec("insert into log_keyword_config(guid,log_keyword_monitor,keyword,regulative,notify_enable,priority,update_time) value (?,?,?,?,?,?,?)", guid.CreateGuid(), param.LogKeywordMonitor, param.Keyword, param.Regulative, param.NotifyEnable, param.Priority, time.Now().Format(models.DatetimeFormat))
 	return
 }
 
 func UpdateLogKeyword(param *models.LogKeywordConfigTable) (err error) {
-	_,err = x.Exec("update log_keyword_config set keyword=?,regulative=?,notify_enable=?,priority=?,update_time=? where guid=?",param.Keyword,param.Regulative,param.NotifyEnable,param.Priority,time.Now().Format(models.DatetimeFormat),param.Guid)
+	_, err = x.Exec("update log_keyword_config set keyword=?,regulative=?,notify_enable=?,priority=?,update_time=? where guid=?", param.Keyword, param.Regulative, param.NotifyEnable, param.Priority, time.Now().Format(models.DatetimeFormat), param.Guid)
 	return
 }
 
 func DeleteLogKeyword(logKeywordConfigGuid string) (err error) {
-	_,err = x.Exec("delete from log_keyword_config where guid=?", logKeywordConfigGuid)
+	_, err = x.Exec("delete from log_keyword_config where guid=?", logKeywordConfigGuid)
 	return
 }
